Came up with a better layout at West MEC
I forgot to push again but I ended up coming up with a better website look here so I'm going to get rid of the code that I have at home and continue with my new code. Almost finished the styles for the pages but need to insert content too.